pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xbcompiler"
version = "0.0.4"
authors = [{ name = "Artemii Kravchuk", email = "artemiykra@gmail.com" }]
description = "Xilia Base Compiler (XLA => LUA)"
readme = "README.md"
requires-python = ">=3.12.4"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
dependencies = ["loguru", "python-magic"]
